Add assert_eq to top-level modules (#3726)
This moves the assert_eq testing functions to top level so that they are
uniformly available as

-   dask.array.assert_eq
-   dask.bag.assert_eq
-   dask.dataframe.assert_eq

THis gets used enough in downstream libraries that it might be nice to
have in a consistent place
