+++ dask/bag/__init__.py
@@ -4,5 +4,6 @@ from .core import (Bag, Item, from_sequence, from_url, to_textfiles, concat,
                    from_delayed, map_partitions, bag_range as range,
                    bag_zip as zip, bag_map as map)
 from .text import read_text
+from .utils import assert_eq
 from ..context import set_options
 from ..base import compute
